start refactoring:
- create mastodon_backup package
- move mastodon-backup.py to mastodon_backup/backup.py
- create argparser for command-line usage
